+++ src/evergreen/__init__.py
@@ -1,5 +1,21 @@
 """Evergreen API Module."""
 from pylibversion import version_tuple_to_str
 
-VERSION = (0, 6, 12)
+VERSION = (0, 6, 13)
 __version__ = version_tuple_to_str(VERSION)
+
+
+# Shortcuts for importing.
+from evergreen.api import EvergreenApi, RetryingEvergreenApi, CachedEvergreenApi, Requester
+from evergreen.build import Build
+from evergreen.commitqueue import CommitQueue
+from evergreen.distro import Distro
+from evergreen.host import Host
+from evergreen.manifest import Manifest
+from evergreen.patch import Patch
+from evergreen.project import Project
+from evergreen.task import Task
+from evergreen.tst import Tst
+from evergreen.stats import TestStats, TaskStats
+from evergreen.task_reliability import TaskReliability
+from evergreen.version import Version