+++ src/evergreen/api.py
@@ -40,7 +40,7 @@ from evergreen.task import Task
 from evergreen.tst import Tst
 from evergreen.stats import TestStats, TaskStats
 from evergreen.task_reliability import TaskReliability
-from evergreen.util import evergreen_input_to_output
+from evergreen.util import evergreen_input_to_output, iterate_by_time_window
 from evergreen.version import Version
 
 structlog.configure(logger_factory=LoggerFactory())
@@ -320,6 +320,22 @@ class _ProjectApi(_BaseEvergreenApi):
         version_list = self._lazy_paginate(url, params)
         return (Version(version, self) for version in version_list)
 
+    def versions_by_project_time_window(self, project_id, start, end,
+                                        requester=Requester.GITTER_REQUEST,
+                                        time_attr='create_time'):
+        """
+        Get an iterator over the patches for the given time window.
+
+        :param project_id: Id of project to query.
+        :param requester: Type of version to query
+        :param start: Timestamp to start returning items (later in time than end).
+        :param end: Timestamp to stop returning items (earlier in time than start).
+        :param time_attr: Attributes to use to window timestamps.
+        :return: Iterator for the given time window.
+        """
+        return iterate_by_time_window(self.versions_by_project(project_id, requester), start, end,
+                                      time_attr)
+
     def patches_by_project(self, project_id, params=None):
         """
         Get a list of patches for the specified project.
@@ -331,6 +347,21 @@ class _ProjectApi(_BaseEvergreenApi):
         url = self._create_url('/projects/{project_id}/patches'.format(project_id=project_id))
         patches = self._lazy_paginate_by_date(url, params)
         return (Patch(patch, self) for patch in patches)
+
+    def patches_by_project_time_window(self, project_id, start, end, params=None,
+                                       time_attr='create_time'):
+        """
+        Get an iterator over the patches for the given time window.
+
+        :param project_id: Id of project to query.
+        :param params: Parameters to pass to endpoint.
+        :param start: Timestamp to start returning items (later in time than end).
+        :param end: Timestamp to stop returning items (earlier in time than start).
+        :param time_attr: Attributes to use to window timestamps.
+        :return: Iterator for the given time window.
+        """
+        return iterate_by_time_window(self.patches_by_project(project_id, params), start, end,
+                                      time_attr)
 
     def commit_queue_for_project(self, project_id):
         """